@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net7.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
     <RootNamespace>Bau.Libraries.LibDbProviders.Odbc</RootNamespace>
 	</PropertyGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="System.Data.Odbc" Version="6.0.1" />
+	  <PackageReference Include="System.Data.Odbc" Version="7.0.0" />
 	</ItemGroup>
 
 	<ItemGroup>
