@@ -1,16 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
+﻿namespace Bau.Libraries.LibDbProviders.Base.Schema;
 
-namespace Bau.Libraries.LibDbProviders.Base.Schema
+/// <summary>
+///		Clase con los datos de una tabla de base de datos
+/// </summary>
+public class TableDbModel : BaseTableDbModel
 {
 	/// <summary>
-	///		Clase con los datos de una tabla de base de datos
+	///		Restricciones
 	/// </summary>
-	public class TableDbModel : BaseTableDbModel
-	{
-		/// <summary>
-		///		Restricciones
-		/// </summary>
-		public List<ConstraintDbModel> Constraints { get; } = new List<ConstraintDbModel>();
-	}
+	public List<ConstraintDbModel> Constraints { get; } = new List<ConstraintDbModel>();
 }
