@@ -63,9 +63,9 @@ namespace Bau.Libraries.LibDbProviders.ODBC
 		/// <summary>
 		///		Obtiene el esquema
 		/// </summary>
-		public override async Task<Base.Schema.SchemaDbModel> GetSchemaAsync(TimeSpan timeout, CancellationToken cancellationToken)
+		public override async Task<Base.Schema.SchemaDbModel> GetSchemaAsync(bool includeSystemTables, TimeSpan timeout, CancellationToken cancellationToken)
 		{
-			return await new OdbcSchemaReader().GetSchemaAsync(this, timeout, cancellationToken);
+			return await new OdbcSchemaReader().GetSchemaAsync(this, includeSystemTables, timeout, cancellationToken);
 		}
 
 		/// <summary>
