@@ -18,7 +18,7 @@ namespace Bau.Libraries.LibDbProviders.Spark.Parser
 		/// <summary>
 		///		Obtiene el esquema
 		/// </summary>
-		internal async Task<SchemaDbModel> GetSchemaAsync(SparkProvider provider, TimeSpan timeout, CancellationToken cancellationToken)
+		internal async Task<SchemaDbModel> GetSchemaAsync(SparkProvider provider, bool includeSystemTables, TimeSpan timeout, CancellationToken cancellationToken)
 		{
 			SchemaDbModel schema = new SchemaDbModel();
 			List<string> tables = new List<string>();
